refactor(xod-client-electron): rename debuggerMiddleware to stopDebuggerOnTabCloseMiddleware
