+++ packages/xod-client-electron/src/debugger/stopDebuggerOnTabCloseMiddleware.js
